+++ Mystiko.Library.Tests.Core/Net/Messages/PeerAnnounceUnitTest.cs
@@ -31,7 +31,7 @@ namespace Mystiko.Library.Tests.Net.Messages
                 1,
                 IPAddress.Loopback,
                 5110,
-                Convert.ToUInt32((DateTime.UtcNow - new DateTime(1970, 01, 01, 0, 0, 0, DateTimeKind.Utc)).TotalSeconds),
+                Convert.ToUInt64((DateTime.UtcNow - new DateTime(1970, 01, 01, 0, 0, 0, DateTimeKind.Utc)).TotalSeconds),
                 new byte[32],
                 new byte[32],
                 0);
@@ -50,7 +50,7 @@ namespace Mystiko.Library.Tests.Net.Messages
                 1,
                 IPAddress.Parse("127.0.0.1"),
                 5110,
-                Convert.ToUInt32((DateTime.UtcNow - new DateTime(1970, 01, 01, 0, 0, 0, DateTimeKind.Utc)).TotalSeconds),
+                Convert.ToUInt64((DateTime.UtcNow - new DateTime(1970, 01, 01, 0, 0, 0, DateTimeKind.Utc)).TotalSeconds),
                 new byte[32],
                 new byte[32],
                 0);
